+++ lib/result_type.dart
@@ -1,3 +1,1 @@
-export 'src/failure.dart';
 export 'src/result.dart';
-export 'src/success.dart';
