+++ pubspec.yaml
@@ -7,8 +7,8 @@ environment:
   sdk: ">=3.0.0 <=4.0.0"
 
 dependencies:
-  meta: ^1.8.0
+  meta: ^1.9.1
 
 dev_dependencies:
-  http: ^0.13.5
-  test: ^1.21.4
+  http: ^1.0.0
+  test: ^1.24.3
